packages/core: split up SignInPage and make it display progress
