+++ packages/core/src/layout/SignInPage/SignInPage.tsx
@@ -14,79 +14,15 @@
  * limitations under the License.
  */
 
-import React, {
-  FC,
-  useLayoutEffect,
-  useState,
-  ComponentType,
-  useMemo,
-} from 'react';
+import React, { FC } from 'react';
 import { Page } from '../Page';
 import { Header } from '../Header';
 import { Content } from '../Content/Content';
 import { ContentHeader } from '../ContentHeader/ContentHeader';
-import { Grid, Typography, Button } from '@material-ui/core';
-import { InfoCard } from '../InfoCard/InfoCard';
-import {
-  SignInPageProps,
-  SignInResult,
-  useApi,
-  configApiRef,
-  useApiHolder,
-  ApiHolder,
-  errorApiRef,
-} from '@backstage/core-api';
-
-const PROVIDER_STORAGE_KEY = '@backstage/core:SignInPage:provider';
-
-type ProviderComponent = ComponentType<SignInPageProps>;
-
-type ProviderLoader = (apis: ApiHolder) => Promise<SignInResult | undefined>;
-
-type SignInProvider = {
-  component: ProviderComponent;
-  loader: ProviderLoader;
-};
-
-const GuestProvider: ProviderComponent = ({ onResult }) => (
-  <Grid item>
-    <InfoCard
-      title="Guest"
-      actions={
-        <Button
-          color="primary"
-          variant="outlined"
-          onClick={() => onResult({ userId: 'guest' })}
-        >
-          Enter
-        </Button>
-      }
-    >
-      <Typography variant="body1">
-        Enter as a Guest User.
-        <br />
-        You will not have a verified identity,
-        <br />
-        meaning some features might be unavailable.
-      </Typography>
-    </InfoCard>
-  </Grid>
-);
-
-const guestLoader: ProviderLoader = async () => {
-  return { userId: 'guest' };
-};
-
-const guestProvider: SignInProvider = {
-  component: GuestProvider,
-  loader: guestLoader,
-};
-
-const signInProviders = {
-  guest: guestProvider,
-};
-
-export type SignInProviderId = keyof typeof signInProviders;
+import { Grid } from '@material-ui/core';
+import { SignInPageProps, useApi, configApiRef } from '@backstage/core-api';
+import { useSignInProviders, SignInProviderId } from './providers';
+import Progress from '../../components/Progress';
 
 export type Props = SignInPageProps & {
   providers: SignInProviderId[];
@@ -94,81 +30,12 @@ export type Props = SignInPageProps & {
 
 export const SignInPage: FC<Props> = ({ onResult, providers }) => {
   const configApi = useApi(configApiRef);
-  const errorApi = useApi(errorApiRef);
-  const apiHolder = useApiHolder();
 
-  // We can't use storageApi here, as it might have a dependency on the IdentityApi
-  const selectedProvider = localStorage.getItem(
-    PROVIDER_STORAGE_KEY,
-  ) as SignInProviderId;
+  const [loading, providerElements] = useSignInProviders(providers, onResult);
 
-  const [attempting, setAttempting] = useState(Boolean(selectedProvider));
-
-  useLayoutEffect(() => {
-    if (!attempting || selectedProvider === null) {
-      return undefined;
-    }
-
-    const provider = signInProviders[selectedProvider];
-    if (!provider) {
-      setAttempting(false);
-      return undefined;
-    }
-
-    let didCancel = false;
-    provider
-      .loader(apiHolder)
-      .then(result => {
-        if (didCancel) {
-          return;
-        }
-        setAttempting(false);
-        if (result) {
-          onResult({
-            ...result,
-            logout: async () => {
-              localStorage.removeItem(PROVIDER_STORAGE_KEY);
-              await result.logout?.();
-            },
-          });
-        }
-      })
-      .catch(error => {
-        if (!didCancel) {
-          errorApi.post(error);
-        }
-      });
-
-    return () => {
-      didCancel = true;
-    };
-  }, [attempting, errorApi, onResult, apiHolder, providers, selectedProvider]);
-
-  const providerElements = useMemo(
-    () =>
-      providers.map(providerId => {
-        const provider = signInProviders[providerId];
-        if (!provider) {
-          throw new Error(`Unknown sign-in provider: ${providerId}`);
-        }
-        const { component: Component } = provider;
-
-        const handleResult = (result: SignInResult) => {
-          localStorage.setItem(PROVIDER_STORAGE_KEY, providerId);
-
-          onResult({
-            ...result,
-            logout: async () => {
-              localStorage.removeItem(PROVIDER_STORAGE_KEY);
-              await result.logout?.();
-            },
-          });
-        };
-
-        return <Component key={providerId} onResult={handleResult} />;
-      }),
-    [providers, onResult],
-  );
+  if (loading) {
+    return <Progress />;
+  }
 
   return (
     <Page>
